Update LGPL3 license that GitHub auto detects (#8)
Following instruction from choosealicense.com

https://choosealicense.com/licenses/lgpl-3.0/
## How to apply this license
This license is an additional set of permissions to the GNU GPLv3 license. Follow the instructions to apply the GNU GPLv3, in the root of your source code. Then add another file named COPYING.LESSER and copy the text.

https://choosealicense.com/licenses/gpl-3.0/
## How to apply this license
Create a text file (typically named COPYING, as per GNU conventions) in the root of your source code and copy the text of the license into the file.